+++ extensions/linear/src/api/initiatives.ts
@@ -1,8 +1,12 @@
-import { Initiative } from "@linear/sdk";
+import { Initiative, Project } from "@linear/sdk";
 import { getLinearClient } from "./linearClient";
 import { sortBy } from "lodash";
 
-export type InitiativeResult = Pick<Initiative, "id" | "name" | "color" | "icon" | "sortOrder" | "description">;
+export type InitiativeResult = Pick<Initiative, "id" | "name" | "color" | "icon" | "sortOrder" | "description"> & {
+  projects?: { nodes: Pick<Project, "id">[] };
+};
+
+type InitiativeList = { initiatives: { nodes: InitiativeResult[] } };
 
 const initiativeFragment = `
   id
@@ -11,14 +15,16 @@ const initiativeFragment = `
   icon
   description
   sortOrder
+  projects {
+    nodes {
+      id
+    }
+  }
 `;
 
 export async function getInitiatives() {
   const { graphQLClient } = getLinearClient();
-  const { data } = await graphQLClient.rawRequest<
-    { initiatives: { nodes: InitiativeResult[] } },
-    Record<string, unknown>
-  >(
+  const { data } = await graphQLClient.rawRequest<InitiativeList, Record<string, unknown>>(
     `
       query {
         initiatives(orderBy: updatedAt) {