+++ extensions/linear/src/components/docs/DocumentActions.tsx
@@ -126,7 +126,7 @@ function DeleteDocument({ doc, mutateDocs, deleteUnsupported }: DocumentActionsP
     return <></>;
   }
 
-  const trash = async () =>
+  const trash = () =>
     confirmAlert({
       title: "Delete Document",
       message: `Are you sure you want to delete '${doc.title}'?`,
@@ -136,7 +136,7 @@ function DeleteDocument({ doc, mutateDocs, deleteUnsupported }: DocumentActionsP
 
   const tryDelete = async () => {
     const toast = await showToast(Toast.Style.Animated, "Deleting document", doc.title);
-    await mutateDocs(deleteDocument(doc.id), {
+    mutateDocs(deleteDocument(doc.id), {
       optimisticUpdate: (data) => {
         if (!data) {
           return undefined;