+++ extensions/moneylover/src/view-wallets.tsx
@@ -6,19 +6,14 @@ import {
   getPreferenceValues,
   showToast,
   Toast,
-  popToRoot,
   Icon,
   Color,
   useNavigation,
 } from '@raycast/api';
-import MoneyLover, { Wallet } from 'moneylover';
-import { useEffect, useState } from 'react';
-import { formatBalance, formatNumber, getBalance } from './utils';
-
-interface Preferences {
-  email: string;
-  password: string;
-}
+import MoneyLover, { AddTransactionParams, Wallet } from 'moneylover';
+import { formatBalance, formatNumber, getBalance, getIcon } from './utils';
+import { FormValidation, useCachedPromise, useForm, usePromise } from '@raycast/utils';
+import { useState } from 'react';
 
 let moneyLoverApi: MoneyLover;
 
@@ -26,62 +21,63 @@ export default function Command() {
   const preferences = getPreferenceValues<Preferences>();
   const { email, password } = preferences;
 
-  const [isLoading, setLoading] = useState(true);
-  const [wallets, setWallets] = useState<Wallet[]>([]);
-
-  useEffect(() => {
-    async function instanciateMoneyLoverClient() {
+  const {
+    isLoading,
+    data: wallets,
+    mutate,
+  } = useCachedPromise(
+    async () => {
+      const toast = await showToast(Toast.Style.Animated, 'Authenticating');
       const accessToken = await MoneyLover.authenticate(email, password);
+      if (!accessToken) throw new Error('Make sure your credentials are valid!');
       moneyLoverApi = new MoneyLover(accessToken);
-    }
 
-    async function fetchWallets() {
+      toast.title = 'Fetching Wallets';
       const wallets = await moneyLoverApi.listWallets();
       const unArchivedWallets = wallets.filter((wallet) => !wallet.archived);
-      setWallets(unArchivedWallets);
+      toast.style = Toast.Style.Success;
+      toast.title = `Fetched ${unArchivedWallets.length} wallets`;
+      return unArchivedWallets;
+    },
+    [],
+    {
+      initialData: [],
     }
-
-    async function main() {
-      setLoading(true);
-
-      try {
-        await instanciateMoneyLoverClient();
-        await fetchWallets();
-      } catch (e) {
-        const err = e as Error;
-        await showToast({
-          title: 'Error',
-          message: err.message,
-          style: Toast.Style.Failure,
-        });
-        await popToRoot();
-      } finally {
-        setLoading(false);
-      }
-    }
-    main();
-  }, []);
-
-  function onBalanceChange(wallet: Wallet) {
-    const updatedWallets = wallets.map((w) => (w._id === wallet._id ? wallet : w));
-    setWallets(updatedWallets);
+  );
+  async function onBalanceChange(wallet: Wallet) {
+    await mutate(undefined, {
+      optimisticUpdate() {
+        const updatedWallets = wallets.map((w) => (w._id === wallet._id ? wallet : w));
+        return updatedWallets;
+      },
+      shouldRevalidateAfter: false,
+    });
   }
 
   return (
-    <List isLoading={isLoading}>
+    <List isLoading={isLoading} searchBarPlaceholder="Search wallet">
       <List.Section title="Wallets">
         {wallets.map((wallet) => (
           <List.Item
             key={wallet._id}
-            icon={`https://static.moneylover.me/img/icon/${wallet.icon}.png`}
+            icon={{ source: getIcon(wallet.icon), fallback: Icon.Wallet }}
             title={wallet.name}
             subtitle={formatBalance(wallet.balance)}
+            accessories={[
+              { icon: wallet.transaction_notification ? Icon.Bell : Icon.BellDisabled },
+              { date: new Date(wallet.updateAt) },
+            ]}
             actions={
               <ActionPanel>
                 <Action.Push
                   icon={{ source: Icon.Pencil, tintColor: Color.Green }}
                   title="Adjust Balance"
                   target={<AdjustBalanceScreen wallet={wallet} onCompleted={onBalanceChange} />}
+                />
+                <Action.Push
+                  icon={Icon.Plus}
+                  title="Add Transaction"
+                  target={<AddTransaction wallet={wallet} onCompleted={onBalanceChange} />}
                 />
               </ActionPanel>
             }
@@ -92,32 +88,25 @@ export default function Command() {
   );
 }
 
-let timeout: NodeJS.Timeout | undefined;
-
 function AdjustBalanceScreen({ wallet, onCompleted }: { wallet: Wallet; onCompleted: (wallet: Wallet) => void }) {
-  const [balance, setBalance] = useState<string>('');
-  const [balanceError, setBalanceError] = useState<string | undefined>();
+  const { itemProps, handleSubmit: formHandleSubmit } = useForm<{ balance: string }>({
+    onSubmit(values) {
+      handleSubmit(values);
+    },
+    validation: {
+      balance(value) {
+        if (!value) return 'The item is required';
+        if (Number.isNaN(+value)) return 'The item must be a number';
+      },
+    },
+  });
 
   const { pop } = useNavigation();
 
   const { currency } = getBalance(wallet.balance);
 
-  function handleOnChange(value: string) {
-    if (Number.isNaN(+value)) {
-      setBalanceError('You must type only numbers');
-      clearTimeout(timeout);
-      timeout = setTimeout(() => {
-        setBalanceError(undefined);
-      }, 2500);
-      return;
-    }
-
-    setBalanceError(undefined);
-    setBalance(value);
-  }
-
   async function handleSubmit(formData: { balance: string }) {
-    await showToast({ title: 'Adjusting balance...', style: Toast.Style.Animated });
+    await showToast({ title: 'Adjusting Balance...', style: Toast.Style.Animated });
 
     try {
       await moneyLoverApi.adjustBalance({
@@ -148,20 +137,135 @@ function AdjustBalanceScreen({ wallet, onCompleted }: { wallet: Wallet; onComple
           <Action.SubmitForm
             icon={{ source: Icon.Checkmark, tintColor: Color.Green }}
             title="Adjust Balance"
+            onSubmit={formHandleSubmit}
+          />
+        </ActionPanel>
+      }
+    >
+      <Form.Description title="Wallet" text={wallet.name} />
+      <Form.TextField title="Balance" placeholder={formatBalance(wallet.balance)} {...itemProps.balance} />
+    </Form>
+  );
+}
+
+function AddTransaction({ wallet, onCompleted }: { wallet: Wallet; onCompleted: (wallet: Wallet) => void }) {
+  const [execute, setExecute] = useState(false);
+
+  type AddTransaction = {
+    walletId: string;
+    amount: string;
+    categoryId: string;
+    displayDate: Date | null;
+    excludeReport?: boolean;
+    image?: string;
+    note?: string;
+    withList?: string[];
+  };
+  const { itemProps, handleSubmit, values } = useForm<AddTransaction>({
+    onSubmit() {
+      setExecute(true);
+    },
+    initialValues: {
+      displayDate: new Date(),
+    },
+    validation: {
+      amount(value) {
+        if (!value) return 'The item is required';
+        if (Number.isNaN(+value)) return 'The item must be a number';
+        if (Number(value) < 1) return 'The item must be greater than 0';
+      },
+      categoryId: FormValidation.Required,
+      displayDate: FormValidation.Required,
+    },
+  });
+
+  const { pop } = useNavigation();
+
+  const { currency, balance } = getBalance(wallet.balance);
+
+  const { isLoading: isAdding } = usePromise(
+    async () => {
+      await showToast({ title: 'Adding Transaction...', style: Toast.Style.Animated });
+
+      const newTransaction: AddTransactionParams = {
+        walletId: wallet._id,
+        amount: +values.amount,
+        categoryId: values.categoryId,
+        displayDate: values.displayDate as Date,
+        excludeReport: values.excludeReport,
+      };
+      if (values.note) newTransaction.note = values.note;
+      const transaction = await moneyLoverApi.addTransaction(newTransaction);
+      if (!transaction) throw new Error();
+
+      wallet.balance[0][currency] = String(balance + +values.amount);
+      await showToast({
+        title: 'Added Transaction',
+        style: Toast.Style.Success,
+      });
+      onCompleted(wallet);
+
+      pop();
+    },
+    [],
+    {
+      execute,
+      async onError() {
+        setExecute(false);
+        await showToast({
+          title: 'Error',
+          message: 'Could not add transaction',
+          style: Toast.Style.Failure,
+        });
+      },
+    }
+  );
+
+  const { isLoading: isFetching, data: categories } = useCachedPromise(
+    async () => {
+      const toast = await showToast(Toast.Style.Animated, 'Fetching Categories');
+      const categories = await moneyLoverApi.listCategories();
+      toast.style = Toast.Style.Success;
+      toast.title = `Fetched ${categories.length} categories`;
+      return categories;
+    },
+    [],
+    {
+      initialData: [],
+    }
+  );
+
+  const isLoading = isFetching || isAdding;
+
+  return (
+    <Form
+      isLoading={isLoading}
+      actions={
+        <ActionPanel>
+          <Action.SubmitForm
+            icon={{ source: Icon.Checkmark, tintColor: Color.Green }}
+            title="Add Transaction"
             onSubmit={handleSubmit}
           />
         </ActionPanel>
       }
     >
-      <Form.TextField
-        id="balance"
-        title="Balance"
-        info={wallet.name}
-        placeholder={formatBalance(wallet.balance)}
-        value={balance}
-        error={balanceError}
-        onChange={handleOnChange}
-      />
+      <Form.Description title="Wallet" text={wallet.name} />
+      <Form.TextField title="Amount" placeholder="100" {...itemProps.amount} />
+      <Form.Dropdown title="Category" {...itemProps.categoryId}>
+        {categories.map((category) => (
+          <Form.Dropdown.Item
+            key={category._id}
+            icon={getIcon(category.icon)}
+            title={category.name}
+            value={category._id}
+          />
+        ))}
+      </Form.Dropdown>
+      <Form.DatePicker title="Display Date" {...itemProps.displayDate} />
+
+      <Form.Checkbox label="Exclude Report" {...itemProps.excludeReport} />
+      <Form.TextField title="Note" {...itemProps.note} />
     </Form>
   );
 }