+++ extensions/namesilo/src/lib/constants.ts
@@ -1,13 +1,14 @@
 import { getPreferenceValues } from "@raycast/api";
 
-export const API_URL = "https://www.namesilo.com/api/";
-const API_KEY = getPreferenceValues<Preferences>().api_key;
+const { api_key, ote_api_key = "", use_ote } = getPreferenceValues<Preferences>();
+export const API_BASE_URL = use_ote ? "https://ote.namesilo.com/" : "https://www.namesilo.com/";
 export const API_PARAMS = {
   version: "1",
   type: "json",
-  key: API_KEY,
+  key: use_ote ? ote_api_key : api_key,
 };
 
 export const NAMESILO_LINKS = {
   search: "https://www.namesilo.com/domain/search-domains",
 };
+export const NAMESILO_IMAGE = "namesilo.png";