+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Linear Changelog
 
+## [Search Projects Fixes] - 2024-08-08
+
+- [#13882](https://github.com/raycast/extensions/issues/13882): Replaced filtering on roadmaps with filtering on initiatives, as initiatives replaced roadmaps.
+- Made Delete Project action alert destructive and added a retry option in case of errors.
+
 ## [Documents and Issues Fixes] - 2024-07-25
 
 - Create Issue: Fixed issue that made projects, cycles, estimates, etc. disappear in the form after consecutive issue creation.